Split store unit tests, use whitebox approach
