+++ internal/pkg/service/buffer/store/file_test.go
@@ -2,15 +2,12 @@ package store
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/keboola/go-client/pkg/keboola"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/filestate"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/key"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/model"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/model/column"
@@ -24,7 +21,7 @@ func TestStore_CreateFile(t *testing.T) {
 	store := newStoreForTest(t)
 	file := newFileForTest()
 
-	_, err := store.createFileOp(ctx, file).Do(ctx, store.client)
+	_, err := store.CreateFileOp(ctx, file).Do(ctx, store.client)
 	assert.NoError(t, err)
 
 	// Check keys
@@ -73,7 +70,7 @@ func TestStore_GetFileOp(t *testing.T) {
 	store := newStoreForTest(t)
 	file := newFileForTest()
 
-	_, err := store.createFileOp(ctx, file).Do(ctx, store.client)
+	_, err := store.CreateFileOp(ctx, file).Do(ctx, store.client)
 	assert.NoError(t, err)
 
 	kv, err := store.getFileOp(ctx, file.FileKey).Do(ctx, store.client)
@@ -117,55 +114,6 @@ file/opened/1000/my-receiver/my-export/2006-01-01T08:04:05.000Z
 }
 >>>>>
 `)
-}
-
-func TestStore_SetFileState_Transitions(t *testing.T) {
-	t.Parallel()
-
-	// Test all transitions
-	testCases := []struct{ from, to filestate.State }{
-		{filestate.Opened, filestate.Closing},
-		{filestate.Closing, filestate.Importing},
-		{filestate.Importing, filestate.Failed},
-		{filestate.Failed, filestate.Importing},
-		{filestate.Importing, filestate.Imported},
-	}
-
-	ctx := context.Background()
-	store := newStoreForTest(t)
-	file := newFileForTest()
-	now, _ := time.Parse(time.RFC3339, "2010-01-01T01:01:01+07:00")
-
-	// Create file
-	_, err := store.createFileOp(ctx, file).Do(ctx, store.client)
-	assert.NoError(t, err)
-
-	for _, tc := range testCases {
-		// Trigger transition
-		ok, err := store.SetFileState(ctx, now, &file, tc.to)
-		desc := fmt.Sprintf("%s -> %s", tc.from, tc.to)
-		assert.NoError(t, err, desc)
-		assert.True(t, ok, desc)
-		assert.Equal(t, tc.to, file.State, desc)
-		expected := `
-<<<<<
-file/<STATE>/1000/my-receiver/my-export/2006-01-01T08:04:05.000Z
------
-%A
-  "state": "<STATE>",%A
-  "<STATE>At": "2009-12-31T18:01:01.000Z"%A
->>>>>
-`
-		etcdhelper.AssertKVsString(t, store.client, strings.ReplaceAll(expected, "<STATE>", tc.to.String()))
-
-		// Test duplicated transition -> nop
-		file.State = tc.from
-		ok, err = store.SetFileState(ctx, time.Now(), &file, tc.to)
-		assert.Error(t, err, desc)
-		assert.Equal(t, fmt.Sprintf(`file "%s" is already in the "%s" state`, file.FileKey, tc.to), err.Error())
-		assert.False(t, ok, desc)
-		assert.Equal(t, tc.to, file.State, desc)
-	}
 }
 
 func newFileForTest() model.File {
